Improved the help
Added a check to hexdump to not allow req_bytes_to_dump to exceed the
  files length.
Updated hexdump to return bytes dumped so we can write more thorough tests.
Removed unneeded clones of file in hexdump as we only need a read
  pointer to the file.
Updated tests for hexdump to include checking the bytes dumped.